pyproject.toml
[tool.poetry]
name = "clever_config"
version = "0.3.0"
description = "Tricky config with features"
repository = "https://github.com/osipov-andrey/python_smart_config"
authors = ["Andrey A. Osipov <developer.osipov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
attributedict = "^0.3.0"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
mypy = "^0.910"
black = "^21.6b0"
flake8 = "^=3.9.2"
pytest-sugar = "^0.9.4"
pytest-clarity = "^1.0.1"
pytest-coverage = "^0.0"
isort = "^5.9.2"
types-requests = "^2.25.0"
flake8-absolute-import = "^1.0"
Faker = "^8.10.1"
types-PyYAML = "^5.4.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py38']

[flake8]
extend-ignore = 'E203'
max-line-length = 120
count = true
statistics = true